requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
azure-cognitiveservices-speech==1.49.1
azure-core==1.39.0
blinker==1.9.0
botbuilder-core==4.17.1
botbuilder-schema==4.17.1
botframework-connector==4.17.1
botframework-streaming==4.17.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
Flask==3.1.3
frozenlist==1.8.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonpickle==3.4.2
MarkupSafe==3.0.3
msal==1.36.0
msrest==0.7.1
multidict==6.7.1
oauthlib==3.3.1
openai==2.32.0
packaging==26.1
propcache==0.4.1
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyJWT==2.12.1
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.33.1
requests-oauthlib==2.0.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
